+++ pytest_asyncio/plugin.py
@@ -5,6 +5,7 @@ import enum
 import functools
 import inspect
 import socket
+import sys
 import warnings
 from asyncio import AbstractEventLoopPolicy
 from textwrap import dedent
@@ -19,21 +20,22 @@ from typing import (
     List,
     Literal,
     Optional,
-    Set,
     TypeVar,
     Union,
     overload,
 )
 
 import pytest
-from _pytest.mark.structures import get_unpacked_marks
 from pytest import (
+    Class,
     Collector,
     Config,
     FixtureRequest,
     Function,
     Item,
     Metafunc,
+    Module,
+    Package,
     Parser,
     PytestCollectionWarning,
     PytestDeprecationWarning,
@@ -200,85 +202,33 @@ def pytest_report_header(config: Config) -> List[str]:
     return [f"asyncio: mode={mode}"]
 
 
-def _preprocess_async_fixtures(
-    collector: Collector,
-    processed_fixturedefs: Set[FixtureDef],
-) -> None:
-    config = collector.config
-    asyncio_mode = _get_asyncio_mode(config)
-    fixturemanager = config.pluginmanager.get_plugin("funcmanage")
-    event_loop_fixture_id = "event_loop"
-    for node, mark in collector.iter_markers_with_node("asyncio_event_loop"):
-        event_loop_fixture_id = node.stash.get(_event_loop_fixture_id, None)
-        if event_loop_fixture_id:
-            break
-    for fixtures in fixturemanager._arg2fixturedefs.values():
-        for fixturedef in fixtures:
-            func = fixturedef.func
-            if fixturedef in processed_fixturedefs or not _is_coroutine_or_asyncgen(
-                func
-            ):
-                continue
-            if not _is_asyncio_fixture_function(func) and asyncio_mode == Mode.STRICT:
-                # Ignore async fixtures without explicit asyncio mark in strict mode
-                # This applies to pytest_trio fixtures, for example
-                continue
-            _make_asyncio_fixture_function(func)
-            function_signature = inspect.signature(func)
-            if "event_loop" in function_signature.parameters:
-                warnings.warn(
-                    PytestDeprecationWarning(
-                        f"{func.__name__} is asynchronous and explicitly "
-                        f'requests the "event_loop" fixture. Asynchronous fixtures and '
-                        f'test functions should use "asyncio.get_running_loop()" '
-                        f"instead."
-                    )
-                )
-            _inject_fixture_argnames(fixturedef, event_loop_fixture_id)
-            _synchronize_async_fixture(fixturedef, event_loop_fixture_id)
-            assert _is_asyncio_fixture_function(fixturedef.func)
-            processed_fixturedefs.add(fixturedef)
-
-
-def _inject_fixture_argnames(
-    fixturedef: FixtureDef, event_loop_fixture_id: str
-) -> None:
+def _inject_fixture_argnames(fixturedef: FixtureDef) -> None:
     """
-    Ensures that `request` and `event_loop` are arguments of the specified fixture.
+    Ensures that `request` is an argument of the specified fixture.
     """
-    to_add = []
-    for name in ("request", event_loop_fixture_id):
-        if name not in fixturedef.argnames:
-            to_add.append(name)
-    if to_add:
-        fixturedef.argnames += tuple(to_add)
+    if "request" not in fixturedef.argnames:
+        fixturedef.argnames += ("request",)
 
 
-def _synchronize_async_fixture(
-    fixturedef: FixtureDef, event_loop_fixture_id: str
-) -> None:
+def _synchronize_async_fixture(fixturedef: FixtureDef) -> None:
     """
     Wraps the fixture function of an async fixture in a synchronous function.
     """
     if inspect.isasyncgenfunction(fixturedef.func):
-        _wrap_asyncgen_fixture(fixturedef, event_loop_fixture_id)
+        _wrap_asyncgen_fixture(fixturedef)
     elif inspect.iscoroutinefunction(fixturedef.func):
-        _wrap_async_fixture(fixturedef, event_loop_fixture_id)
+        _wrap_async_fixture(fixturedef)
 
 
 def _add_kwargs(
     func: Callable[..., Any],
     kwargs: Dict[str, Any],
-    event_loop_fixture_id: str,
-    event_loop: asyncio.AbstractEventLoop,
     request: SubRequest,
 ) -> Dict[str, Any]:
     sig = inspect.signature(func)
     ret = kwargs.copy()
     if "request" in sig.parameters:
         ret["request"] = request
-    if event_loop_fixture_id in sig.parameters:
-        ret[event_loop_fixture_id] = event_loop
     return ret
 
 
@@ -301,7 +251,7 @@ def _perhaps_rebind_fixture_func(
     return func
 
 
-def _wrap_asyncgen_fixture(fixturedef: FixtureDef, event_loop_fixture_id: str) -> None:
+def _wrap_asyncgen_fixture(fixturedef: FixtureDef) -> None:
     fixture = fixturedef.func
 
     @functools.wraps(fixture)
@@ -309,10 +259,8 @@ def _wrap_asyncgen_fixture(fixturedef: FixtureDef, event_loop_fixture_id: str) -
         func = _perhaps_rebind_fixture_func(
             fixture, request.instance, fixturedef.unittest
         )
-        event_loop = kwargs.pop(event_loop_fixture_id)
-        gen_obj = func(
-            **_add_kwargs(func, kwargs, event_loop_fixture_id, event_loop, request)
-        )
+        event_loop = asyncio.get_event_loop()
+        gen_obj = func(**_add_kwargs(func, kwargs, request))
 
         async def setup():
             res = await gen_obj.__anext__()
@@ -340,7 +288,7 @@ def _wrap_asyncgen_fixture(fixturedef: FixtureDef, event_loop_fixture_id: str) -
     fixturedef.func = _asyncgen_fixture_wrapper
 
 
-def _wrap_async_fixture(fixturedef: FixtureDef, event_loop_fixture_id: str) -> None:
+def _wrap_async_fixture(fixturedef: FixtureDef) -> None:
     fixture = fixturedef.func
 
     @functools.wraps(fixture)
@@ -348,15 +296,12 @@ def _wrap_async_fixture(fixturedef: FixtureDef, event_loop_fixture_id: str) -> N
         func = _perhaps_rebind_fixture_func(
             fixture, request.instance, fixturedef.unittest
         )
-        event_loop = kwargs.pop(event_loop_fixture_id)
 
         async def setup():
-            res = await func(
-                **_add_kwargs(func, kwargs, event_loop_fixture_id, event_loop, request)
-            )
+            res = await func(**_add_kwargs(func, kwargs, request))
             return res
 
-        return event_loop.run_until_complete(setup())
+        return asyncio.get_event_loop().run_until_complete(setup())
 
     fixturedef.func = _async_fixture_wrapper
 
@@ -492,24 +437,6 @@ class AsyncHypothesisTest(PytestAsyncioFunction):
         super().runtest()
 
 
-_HOLDER: Set[FixtureDef] = set()
-
-
-# The function name needs to start with "pytest_"
-# see https://github.com/pytest-dev/pytest/issues/11307
-@pytest.hookimpl(specname="pytest_pycollect_makeitem", tryfirst=True)
-def pytest_pycollect_makeitem_preprocess_async_fixtures(
-    collector: Union[pytest.Module, pytest.Class], name: str, obj: object
-) -> Union[
-    pytest.Item, pytest.Collector, List[Union[pytest.Item, pytest.Collector]], None
-]:
-    """A pytest hook to collect asyncio coroutines."""
-    if not collector.funcnamefilter(name):
-        return None
-    _preprocess_async_fixtures(collector, _HOLDER)
-    return None
-
-
 # The function name needs to start with "pytest_"
 # see https://github.com/pytest-dev/pytest/issues/11307
 @pytest.hookimpl(specname="pytest_pycollect_makeitem", hookwrapper=True)
@@ -542,54 +469,78 @@ def pytest_pycollect_makeitem_convert_async_functions_to_subclass(
 
 
 _event_loop_fixture_id = StashKey[str]
+_fixture_scope_by_collector_type = {
+    Class: "class",
+    Module: "module",
+    Package: "package",
+    Session: "session",
+}
 
 
 @pytest.hookimpl
 def pytest_collectstart(collector: pytest.Collector):
-    if not isinstance(collector, (pytest.Class, pytest.Module)):
-        return
-    # pytest.Collector.own_markers is empty at this point,
-    # so we rely on _pytest.mark.structures.get_unpacked_marks
-    marks = get_unpacked_marks(collector.obj)
-    for mark in marks:
-        if not mark.name == "asyncio_event_loop":
-            continue
-
-        # There seem to be issues when a fixture is shadowed by another fixture
-        # and both differ in their params.
-        # https://github.com/pytest-dev/pytest/issues/2043
-        # https://github.com/pytest-dev/pytest/issues/11350
-        # As such, we assign a unique name for each event_loop fixture.
-        # The fixture name is stored in the collector's Stash, so it can
-        # be injected when setting up the test
-        event_loop_fixture_id = f"{collector.nodeid}::<event_loop>"
+    # Session is not a PyCollector type, so it doesn't have a corresponding
+    # "obj" attribute to attach a dynamic fixture function to.
+    # However, there's only one session per pytest run, so there's no need to
+    # create the fixture dynamically. We can simply define a session-scoped
+    # event loop fixture once in the plugin code.
+    if isinstance(collector, Session):
+        event_loop_fixture_id = _session_event_loop.__name__
         collector.stash[_event_loop_fixture_id] = event_loop_fixture_id
+        return
+    if not isinstance(collector, (Class, Module, Package)):
+        return
+    # There seem to be issues when a fixture is shadowed by another fixture
+    # and both differ in their params.
+    # https://github.com/pytest-dev/pytest/issues/2043
+    # https://github.com/pytest-dev/pytest/issues/11350
+    # As such, we assign a unique name for each event_loop fixture.
+    # The fixture name is stored in the collector's Stash, so it can
+    # be injected when setting up the test
+    event_loop_fixture_id = f"{collector.nodeid}::<event_loop>"
+    collector.stash[_event_loop_fixture_id] = event_loop_fixture_id
 
-        @pytest.fixture(
-            scope="class" if isinstance(collector, pytest.Class) else "module",
-            name=event_loop_fixture_id,
-        )
-        def scoped_event_loop(
-            *args,  # Function needs to accept "cls" when collected by pytest.Class
-            event_loop_policy,
-        ) -> Iterator[asyncio.AbstractEventLoop]:
-            new_loop_policy = event_loop_policy
-            old_loop_policy = asyncio.get_event_loop_policy()
-            old_loop = asyncio.get_event_loop()
-            asyncio.set_event_loop_policy(new_loop_policy)
-            loop = asyncio.new_event_loop()
-            asyncio.set_event_loop(loop)
-            yield loop
-            loop.close()
-            asyncio.set_event_loop_policy(old_loop_policy)
-            asyncio.set_event_loop(old_loop)
+    @pytest.fixture(
+        scope=_fixture_scope_by_collector_type[type(collector)],
+        name=event_loop_fixture_id,
+    )
+    def scoped_event_loop(
+        *args,  # Function needs to accept "cls" when collected by pytest.Class
+        event_loop_policy,
+    ) -> Iterator[asyncio.AbstractEventLoop]:
+        new_loop_policy = event_loop_policy
+        old_loop_policy = asyncio.get_event_loop_policy()
+        old_loop = asyncio.get_event_loop()
+        asyncio.set_event_loop_policy(new_loop_policy)
+        loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(loop)
+        yield loop
+        loop.close()
+        asyncio.set_event_loop_policy(old_loop_policy)
+        asyncio.set_event_loop(old_loop)
 
-        # @pytest.fixture does not register the fixture anywhere, so pytest doesn't
-        # know it exists. We work around this by attaching the fixture function to the
-        # collected Python class, where it will be picked up by pytest.Class.collect()
-        # or pytest.Module.collect(), respectively
-        collector.obj.__pytest_asyncio_scoped_event_loop = scoped_event_loop
-        break
+    # @pytest.fixture does not register the fixture anywhere, so pytest doesn't
+    # know it exists. We work around this by attaching the fixture function to the
+    # collected Python class, where it will be picked up by pytest.Class.collect()
+    # or pytest.Module.collect(), respectively
+    collector.obj.__pytest_asyncio_scoped_event_loop = scoped_event_loop
+    # When collector is a package, collector.obj is the package's __init__.py.
+    # pytest doesn't seem to collect fixtures in __init__.py.
+    # Using parsefactories to collect fixtures in __init__.py their baseid will end
+    # with "__init__.py", thus limiting the scope of the fixture to the init module.
+    # Therefore, we tell the pluginmanager explicitly to collect the fixtures
+    # in the init module, but strip "__init__.py" from the baseid
+    # Possibly related to https://github.com/pytest-dev/pytest/issues/4085
+    if isinstance(collector, Package):
+        fixturemanager = collector.config.pluginmanager.get_plugin("funcmanage")
+        package_node_id = _removesuffix(collector.nodeid, "__init__.py")
+        fixturemanager.parsefactories(collector.obj, nodeid=package_node_id)
+
+
+def _removesuffix(s: str, suffix: str) -> str:
+    if sys.version_info < (3, 9):
+        return s[: -len(suffix)]
+    return s.removesuffix(suffix)
 
 
 def pytest_collection_modifyitems(
@@ -608,7 +559,9 @@ def pytest_collection_modifyitems(
     if _get_asyncio_mode(config) != Mode.AUTO:
         return
     for item in items:
-        if isinstance(item, PytestAsyncioFunction):
+        if isinstance(item, PytestAsyncioFunction) and not item.get_closest_marker(
+            "asyncio"
+        ):
             item.add_marker("asyncio")
 
 
@@ -626,32 +579,34 @@ _REDEFINED_EVENT_LOOP_FIXTURE_WARNING = dedent(
 
 @pytest.hookimpl(tryfirst=True)
 def pytest_generate_tests(metafunc: Metafunc) -> None:
-    for event_loop_provider_node, _ in metafunc.definition.iter_markers_with_node(
-        "asyncio_event_loop"
-    ):
-        event_loop_fixture_id = event_loop_provider_node.stash.get(
-            _event_loop_fixture_id, None
-        )
-        if event_loop_fixture_id:
-            # This specific fixture name may already be in metafunc.argnames, if this
-            # test indirectly depends on the fixture. For example, this is the case
-            # when the test depends on an async fixture, both of which share the same
-            # asyncio_event_loop mark.
-            if event_loop_fixture_id in metafunc.fixturenames:
-                continue
-            fixturemanager = metafunc.config.pluginmanager.get_plugin("funcmanage")
-            if "event_loop" in metafunc.fixturenames:
-                raise MultipleEventLoopsRequestedError(
-                    _MULTIPLE_LOOPS_REQUESTED_ERROR
-                    % (metafunc.definition.nodeid, event_loop_provider_node.nodeid),
-                )
-            # Add the scoped event loop fixture to Metafunc's list of fixture names and
-            # fixturedefs and leave the actual parametrization to pytest
-            metafunc.fixturenames.insert(0, event_loop_fixture_id)
-            metafunc._arg2fixturedefs[
-                event_loop_fixture_id
-            ] = fixturemanager._arg2fixturedefs[event_loop_fixture_id]
-            break
+    marker = metafunc.definition.get_closest_marker("asyncio")
+    if not marker:
+        return
+    scope = marker.kwargs.get("scope", "function")
+    if scope == "function":
+        return
+    event_loop_node = _retrieve_scope_root(metafunc.definition, scope)
+    event_loop_fixture_id = event_loop_node.stash.get(_event_loop_fixture_id, None)
+
+    if event_loop_fixture_id:
+        # This specific fixture name may already be in metafunc.argnames, if this
+        # test indirectly depends on the fixture. For example, this is the case
+        # when the test depends on an async fixture, both of which share the same
+        # asyncio_event_loop mark.
+        if event_loop_fixture_id in metafunc.fixturenames:
+            return
+        fixturemanager = metafunc.config.pluginmanager.get_plugin("funcmanage")
+        if "event_loop" in metafunc.fixturenames:
+            raise MultipleEventLoopsRequestedError(
+                _MULTIPLE_LOOPS_REQUESTED_ERROR
+                % (metafunc.definition.nodeid, event_loop_node.nodeid),
+            )
+        # Add the scoped event loop fixture to Metafunc's list of fixture names and
+        # fixturedefs and leave the actual parametrization to pytest
+        metafunc.fixturenames.insert(0, event_loop_fixture_id)
+        metafunc._arg2fixturedefs[
+            event_loop_fixture_id
+        ] = fixturemanager._arg2fixturedefs[event_loop_fixture_id]
 
 
 @pytest.hookimpl(hookwrapper=True)
@@ -659,6 +614,7 @@ def pytest_fixture_setup(
     fixturedef: FixtureDef, request: SubRequest
 ) -> Optional[object]:
     """Adjust the event loop policy when an event loop is produced."""
+    async_mode = _get_asyncio_mode(request.config)
     if fixturedef.argname == "event_loop":
         # The use of a fixture finalizer is preferred over the
         # pytest_fixture_post_finalizer hook. The fixture finalizer is invoked once
@@ -697,7 +653,23 @@ def pytest_fixture_setup(
             pass
         policy.set_event_loop(loop)
         return
-
+    elif inspect.iscoroutinefunction(fixturedef.func) or inspect.isasyncgenfunction(
+        fixturedef.func
+    ):
+        if async_mode == Mode.STRICT and not _is_asyncio_fixture_function(
+            fixturedef.func
+        ):
+            yield
+            return
+        if "event_loop" in inspect.signature(fixturedef.func).parameters:
+            raise pytest.UsageError(
+                f"{fixturedef.func.__name__} is asynchronous and explicitly "
+                f'requests the "event_loop" fixture. Asynchronous fixtures and '
+                f'test functions should use "asyncio.get_running_loop()" '
+                f"instead."
+            )
+        _inject_fixture_argnames(fixturedef)
+        _synchronize_async_fixture(fixturedef)
     yield
 
 
@@ -844,11 +816,12 @@ def pytest_runtest_setup(item: pytest.Item) -> None:
     marker = item.get_closest_marker("asyncio")
     if marker is None:
         return
-    event_loop_fixture_id = "event_loop"
-    for node, mark in item.iter_markers_with_node("asyncio_event_loop"):
-        event_loop_fixture_id = node.stash.get(_event_loop_fixture_id, None)
-        if event_loop_fixture_id:
-            break
+    scope = marker.kwargs.get("scope", "function")
+    if scope != "function":
+        parent_node = _retrieve_scope_root(item, scope)
+        event_loop_fixture_id = parent_node.stash[_event_loop_fixture_id]
+    else:
+        event_loop_fixture_id = "event_loop"
     fixturenames = item.fixturenames  # type: ignore[attr-defined]
     # inject an event loop fixture for all async tests
     if "event_loop" in fixturenames:
@@ -862,6 +835,24 @@ def pytest_runtest_setup(item: pytest.Item) -> None:
             "test function `%r` is using Hypothesis, but pytest-asyncio "
             "only works with Hypothesis 3.64.0 or later." % item
         )
+
+
+def _retrieve_scope_root(item: Union[Collector, Item], scope: str) -> Collector:
+    node_type_by_scope = {
+        "class": Class,
+        "module": Module,
+        "package": Package,
+        "session": Session,
+    }
+    scope_root_type = node_type_by_scope[scope]
+    for node in reversed(item.listchain()):
+        if isinstance(node, scope_root_type):
+            return node
+    error_message = (
+        f"{item.name} is marked to be run in an event loop with scope {scope}, "
+        f"but is not part of any {scope}."
+    )
+    raise pytest.UsageError(error_message)
 
 
 @pytest.fixture
@@ -878,6 +869,22 @@ def event_loop(request: FixtureRequest) -> Iterator[asyncio.AbstractEventLoop]:
     loop.__original_fixture_loop = True  # type: ignore[attr-defined]
     yield loop
     loop.close()
+
+
+@pytest.fixture(scope="session")
+def _session_event_loop(
+    request: FixtureRequest, event_loop_policy: AbstractEventLoopPolicy
+) -> Iterator[asyncio.AbstractEventLoop]:
+    new_loop_policy = event_loop_policy
+    old_loop_policy = asyncio.get_event_loop_policy()
+    old_loop = asyncio.get_event_loop()
+    asyncio.set_event_loop_policy(new_loop_policy)
+    loop = asyncio.new_event_loop()
+    asyncio.set_event_loop(loop)
+    yield loop
+    loop.close()
+    asyncio.set_event_loop_policy(old_loop_policy)
+    asyncio.set_event_loop(old_loop)
 
 
 @pytest.fixture(scope="session", autouse=True)