+++ docs/source/reference/changelog.rst
@@ -2,13 +2,24 @@
 Changelog
 =========
 
+1.0.0 (UNRELEASED)
+==================
+- BREAKING: Asynchronous fixtures can no longer request the *event_loop* fixture
+- BREAKING: Parametrizations and custom implementations of the *event_loop* fixture no longer have any effect on async fixtures
+
 0.23.0 (UNRELEASED)
 ===================
-- Removes pytest-trio from the test dependencies `#620 <https://github.com/pytest-dev/pytest-asyncio/pull/620>`_
+This release is backwards-compatible with v0.21.
+Changes are non-breaking, unless you upgrade from v0.22.
+
+- BREAKING: The *asyncio_event_loop* mark has been removed. Event loops with class, module, package, and session scopes can be requested via the *scope* keyword argument to the _asyncio_ mark.
 - Introduces the *event_loop_policy* fixture which allows testing with non-default or multiple event loops  `#662 <https://github.com/pytest-dev/pytest-asyncio/pull/662>`_
+- Removes pytest-trio from the test dependencies `#620 <https://github.com/pytest-dev/pytest-asyncio/pull/620>`_
 
 0.22.0 (2023-10-31)
 ===================
+This release has been yanked from PyPI due to fundamental issues with the _asyncio_event_loop_ mark.
+
 - Class-scoped and module-scoped event loops can be requested
   via the _asyncio_event_loop_ mark. `#620 <https://github.com/pytest-dev/pytest-asyncio/pull/620>`_
 - Deprecate redefinition of the `event_loop` fixture. `#587 <https://github.com/pytest-dev/pytest-asyncio/issues/531>`_